requirements.txt
# Core LangChain packages
langchain
langchain-community
langchain-huggingface

# HuggingFace transformers for local models
transformers
sentence-transformers
torch

# Vector database
chromadb

# Document loading
unstructured

# Environment variables
python-dotenv

# Note: After installing requirements, run: pip install "unstructured[md]" for markdown support 
